UCP: Rename files - stub_ep.c -> wireup_ep.c
