+++ src/ucp/core/ucp_ep.c
@@ -9,7 +9,7 @@
 #include "ucp_ep.inl"
 #include "ucp_request.inl"
 
-#include <ucp/wireup/stub_ep.h>
+#include <ucp/wireup/wireup_ep.h>
 #include <ucp/wireup/wireup.h>
 #include <ucp/tag/eager.h>
 #include <ucp/tag/offload.h>
